+++ src/modules/feature/board-file-element/shared/types/file-interaction-type.ts
